+++ src/psf_generator/__init__.py
@@ -1,6 +1,8 @@
 # Copyright Biomedical Imaging Group, EPFL 2025
 __version__ = '0.2.0'
 
+from .imaging import CartesianDipoleImager, SphericalDipoleImager
+from .modalities import COBRIMicroscope, DarkFieldMicroscope, ISCATMicroscope, Particle
 from .propagators import (
     ScalarCartesianPropagator,
     ScalarSphericalPropagator,
@@ -13,4 +15,10 @@ __all__ = [
     'ScalarSphericalPropagator',
     'VectorialCartesianPropagator',
     'VectorialSphericalPropagator',
+    'SphericalDipoleImager',
+    'CartesianDipoleImager',
+    'Particle',
+    'ISCATMicroscope',
+    'COBRIMicroscope',
+    'DarkFieldMicroscope',
 ]